requirements.txt
Django==2.0.5
djangorestframework==3.8.2
mysqlclient==1.3.12
pytz==2018.4
wheel==0.24.0
